requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
discord.py==2.6.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pyfiglet==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
urllib3==2.5.0
yarl==1.20.1
